refactor: clean up dead code and fix integration test imports
- Remove duplicate methods: set_window_position(), reload(), resize(), move_to()

- Add #[allow(dead_code)] with comments for API methods: is_window_valid(), toggle_fullscreen(), hide(), focus()

- Fix integration test imports: use _core:: instead of auroraview::

- Add missing batch_interval_ms field to MessageQueueConfig in tests

- Use ..Default::default() for WebViewConfig in tests

- Fix boolean expression simplification in warmup.rs

+++ src/webview/backend/native.rs
@@ -643,7 +643,7 @@ impl NativeBackend {
                     // For Container mode (Qt integration), apply anti-flicker optimizations
                     // This prevents the window from being visible before Qt's createWindowContainer
                     if matches!(config.embed_mode, EmbedMode::Container) {
-                        Self::apply_anti_flicker_optimizations(hwnd_value as isize);
+                        Self::apply_anti_flicker_optimizations(hwnd_value);
                     }
                 }
             }
